Refactor: 채팅방 상단 캐릭터 정보 컴포넌트 분리
